% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_hands.R
\docType{data}
\name{pair_ranks}
\alias{pair_ranks}
\title{pair_ranks}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2652 rows and 12 columns.}
\usage{
pair_ranks
}
\description{
pair_ranks
}
\keyword{datasets}
