% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairranks}
\alias{pairranks}
\title{pairranks}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2652 rows and 12 columns.}
\usage{
pairranks
}
\description{
pairranks
}
\keyword{datasets}
