% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_admin.R
\docType{data}
\name{poker_admin}
\alias{poker_admin}
\title{poker_admin}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
poker_admin
}
\description{
poker_admin
}
\keyword{datasets}
