% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_session.R
\docType{data}
\name{poker_session}
\alias{poker_session}
\title{poker_session}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
poker_session
}
\description{
poker_session
}
\keyword{datasets}
