% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tournament.R
\docType{data}
\name{tournament}
\alias{tournament}
\title{poker tournament}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
tournament
}
\description{
poker tournament
}
\keyword{datasets}
