% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_hands.R
\docType{data}
\name{value_dict}
\alias{value_dict}
\title{value_dict}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13 rows and 2 columns.}
\source{
me
}
\usage{
value_dict
}
\description{
value_dict
}
\keyword{datasets}
